update activeData tests red
