+++ cypress/integration/backend/Routes/index.spec.ts
@@ -84,32 +84,6 @@ describe('API /user', () => {
     });
   });
 
-  it("Should post a user's activeData", () => {
-    const activeData: ActiveUserInterface = {
-      // set by server
-      userId: '61bd4e30c59b1240d579b0b3', // RodrygoG
-      username: 'RodrygoG',
-      profileLink: 'someProfileLink',
-
-      // sent by client
-      location: {coordinates: {latitude: -23.5, longitude: -46.6}},
-      music: {
-        trackName: 'The Rover',
-        authorsNames: ['Led Zepellin'],
-        coverImage: {url: 'https://i.scdn.co/image/ab67616d0000b273f8f9b9b9b9b9b9b9b9b9b9b9b9', width: 300, height: 300},
-        status: 'playing',
-      },
-    };
-
-    cy.request('POST', '/user/active-data', {
-      activeData
-    }).then(res => {
-      expect(res.status).to.eq(200);
-      expect(res.body.message).to.eq('Active data has been set');
-    });
-  });
-
-
   it('It should retrieve near users as an array', () => {
     // With user logged in
     cy.request('/near-users', {id: '61bd4e30c59b1240d579b0b3'})
